Create HasUpdateIdentifier Trait and implement it on Article and Proceeding repository
